rename: common 폴더 내부로 공통 컴포넌트 이동
