+++ src/components/@common/Button/index.tsx
@@ -1,4 +1,4 @@
-import Typography from 'components/Typography';
+import Typography from 'components/@common/Typography';
 import {
   StyleSheet,
   TouchableOpacity,
